backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

langchain==0.1.20
langchain-community==0.0.38


httpx==0.27.0
pydantic==1.10.13

pgvector==0.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

python-multipart==0.0.6

sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1

